+++ src/main/java/nextstep/subway/station/dto/StationResponse.java
@@ -1,23 +1,28 @@
-package nextstep.subway.dto;
+package nextstep.subway.station.dto;
 
-import nextstep.subway.domain.Station;
+import nextstep.subway.station.domain.Station;
 
 import java.time.LocalDateTime;
 
 public class StationResponse {
-    private Long id;
-    private String name;
-    private LocalDateTime createdDate;
-    private LocalDateTime modifiedDate;
+    private final Long id;
+    private final String name;
+    private final LocalDateTime createdDate;
+    private final LocalDateTime modifiedDate;
 
-    public static StationResponse of(Station station) {
+    public static StationResponse of(final Station station) {
         return new StationResponse(station.getId(), station.getName(), station.getCreatedDate(), station.getModifiedDate());
     }
 
     public StationResponse() {
+        this.id = null;
+        this.name = null;
+        this.createdDate = null;
+        this.modifiedDate = null;
+
     }
 
-    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    public StationResponse(final Long id, final String name, final LocalDateTime createdDate, final LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
         this.createdDate = createdDate;
