+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -3,9 +3,9 @@ package nextstep.subway.station;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.domain.Station;
-import nextstep.subway.dto.StationRequest;
-import nextstep.subway.util.DatabaseCleanup;
+import nextstep.subway.common.domain.util.DatabaseCleanup;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.dto.StationRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,8 +16,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
